Added soem methods for inventory.js
