+++ Frontend_Code/html/main.html
@@ -240,6 +240,8 @@
         </li>
       </ul>
     </nav>
-    <script src="/Frontend_Code/js/script.js"></script>
+    <!-- Script to access the supabase API -->
+    <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js@2"></script>
+    <script type="module" src="/Frontend_Code/js/script.js"></script>
   </body>
 </html>
